+++ packages/components-vue/projects/vue-wrapper/src/public-api.ts
@@ -4,4 +4,3 @@ export * from './lib/types';
 export { useToastManager } from './utils';
 export { default as PorscheDesignSystemProvider } from './PorscheDesignSystemProvider.vue';
 export { createPorscheDesignSystem, usePorscheDesignSystemPlugin } from './plugin';
-